[Feat] 공통 응답 구조 생성
