--- src/main/java/domain/product/repository/ProductRepository.java
@@ -1,7 +1,0 @@
-package domain.product.repository;
-
-import domain.product.model.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
