--- src/main/java/domain/user/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package domain.user.repository;
-
-import domain.user.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
