+++ src/main/java/org/sopt/SOPT_36_COLLABORATION_SERVER_TEMU/domain/product/controller/ProductController.java
@@ -1,6 +1,6 @@
-package domain.product.controller;
+package org.sopt.SOPT_36_COLLABORATION_SERVER_TEMU.domain.product.controller;
 
-import domain.product.service.ProductService;
+import org.sopt.SOPT_36_COLLABORATION_SERVER_TEMU.domain.product.service.ProductService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
