+++ src/main/java/org/sopt/SOPT_36_COLLABORATION_SERVER_TEMU/domain/product/model/Product.java
@@ -1,4 +1,4 @@
-package domain.product.model;
+package org.sopt.SOPT_36_COLLABORATION_SERVER_TEMU.domain.product.model;
 
 import jakarta.persistence.*;
 import lombok.*;
