+++ src/main/java/org/sopt/SOPT_36_COLLABORATION_SERVER_TEMU/domain/product/model/ProductColor.java
@@ -1,0 +1,22 @@
+package org.sopt.SOPT_36_COLLABORATION_SERVER_TEMU.domain.product.model;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Setter
+public class ProductColor {
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "productColorId")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "productId", nullable = false)
+    private Product product;
+
+    @Column(nullable = false)
+    private String color;
+}
