+++ src/main/java/org/sopt/SOPT_36_COLLABORATION_SERVER_TEMU/domain/product/model/ProductDetail.java
@@ -1,16 +1,16 @@
-package domain.product.model;
+package org.sopt.SOPT_36_COLLABORATION_SERVER_TEMU.domain.product.model;
 
 import jakarta.persistence.*;
 import lombok.*;
 
 @Entity
 @Getter
-@AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
 @Setter
-public class ProductColor {
+public class ProductDetail {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "productColorId")
+    @Column(name = "productDetailId")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
@@ -18,5 +18,5 @@ public class ProductColor {
     private Product product;
 
     @Column(nullable = false)
-    private String color;
+    private String imageUrl;
 }
